+++ platform/mac/framework/AGDnsProxy.h
@@ -113,6 +113,11 @@ typedef NS_ENUM(NSInteger, AGDnsBlockingMode) {
      * an empty SOA response if request type is not A/AAAA.
      */
     AGBM_ADDRESS,
+    /**
+     * Respond with an address that is all zeroes regardless of the custom blocking address setting,
+     * or an empty SOA response if request type is not A/AAAA.
+     */
+    AGBM_UNSPECIFIED_ADDRESS,
 };
 
 /**