+++ proxy/dnsproxy.cpp
@@ -1,10 +1,19 @@
-#include <algorithm>
-
-#include "common/logger.h"
-#include "dns/common/version.h"
 #include "dns/proxy/dnsproxy.h"
 
-#include "dns64.h"
+#include <memory>
+#include <optional>
+#include <utility>
+#include <vector>
+
+#include "common/coro.h"
+#include "common/defs.h"
+#include "common/logger.h"
+#include "dns/common/dns_utils.h"
+#include "dns/common/event_loop.h"
+#include "dns/common/version.h"
+#include "dns/proxy/dnsproxy_events.h"
+#include "dns/proxy/dnsproxy_settings.h"
+#include "dns/upstream/upstream.h"
 #include "dns_forwarder.h"
 #include "dnsproxy_listener.h"
 
@@ -70,7 +79,13 @@ static const DnsProxySettings DEFAULT_PROXY_SETTINGS = {
         .outbound_proxy = std::nullopt,
         .block_ipv6 = false,
         .ipv6_available = true,
+#ifdef _WIN32
+        // On Windows, a funky RCODE leads to issues with the system resolver
+        // trying other servers and AdGuard VPN blocking those requests.
+        .adblock_rules_blocking_mode = DnsProxyBlockingMode::UNSPECIFIED_ADDRESS,
+#else
         .adblock_rules_blocking_mode = DnsProxyBlockingMode::REFUSED,
+#endif
         .hosts_rules_blocking_mode = DnsProxyBlockingMode::ADDRESS,
         .dns_cache_size = 1000,
         .optimistic_cache = true,